+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.3.0-beta.3 (Unreleased)
+## 1.4.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,14 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.3.0 (2024-06-07)
+
+### Other Changes
+
+* Changed `AzureMonitorLogExporter` to be internal.
+  This will be changed back to public in our next Beta while we experiment with options to enable log filtering.
+  ([#44479](https://github.com/Azure/azure-sdk-for-net/pull/44479))
 
 ## 1.3.0-beta.2 (2024-05-08)
 
@@ -27,8 +35,8 @@
 ### Other Changes
 
 * Update OpenTelemetry dependencies
-  ([#43197](https://github.com/Azure/azure-sdk-for-net/pull/43197))
-  - OpenTelemetry 1.8.0
+  ([#43688](https://github.com/Azure/azure-sdk-for-net/pull/43688))
+  - OpenTelemetry 1.8.1
 
 ## 1.3.0-beta.1 (2024-02-08)