+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.19.0-beta.1 (Unreleased)
+## 12.19.0-beta.1 (2024-06-11)
 - Added support for service version 2024-08-04.
 - Added more detailed messaging for authorization failure cases.
 - Added support for snapshot management on NFS shares.